Fix and Add Deletion of ToDos

+++ app.js
@@ -98,16 +98,28 @@ app.post("/", (req, res) => {
 });
 
 app.post("/delete", (req, res) => {
-    console.log(req.body.checkbox);
     const checkedItemId = req.body.checkbox;
-    Item.findByIdAndRemove(checkedItemId, (err) => {
-        if (err) {
-            console.log(err);
-        } else {
-            console.log("Successfully Removed!");
-            res.redirect("/")
-        }
-    });
+    const listName = req.body.listName
+    console.log(listName);
+
+    if (listName === date.getDate()) {
+        Item.findByIdAndRemove(checkedItemId, (err) => {
+            if (err) {
+                console.log(err);
+            } else {
+                console.log("Successfully Removed!");
+                res.redirect("/")
+            }
+        });
+    } else {
+        List.findOneAndUpdate({name: listName}, {$pull: {items: {_id: checkedItemId}}}, (err, foundList) => {
+            if (!err) {
+                res.redirect("/" + listName);
+            }
+        });
+    }
+
+    
 });
 
 // app.get("/work", (req, res) => {
